ws: refactor ws_llc_set_mode_switch()
WI_SUN-3223

+++ 6lbr/ws/ws_llc.h
@@ -123,7 +123,8 @@ void ws_llc_timer_seconds(struct net_if *interface, uint16_t seconds_update);
 bool ws_llc_eapol_relay_forward_filter(struct net_if *interface, const uint8_t *joiner_eui64,
                                        uint8_t mac_sequency, uint64_t rx_timestamp);
 
-int8_t ws_llc_set_mode_switch(struct net_if *interface, int mode, uint8_t phy_mode_id, uint8_t *neighbor_mac_address);
+int8_t ws_llc_set_mode_switch(struct net_if *interface, uint8_t mode, uint8_t phy_mode_id,
+                              uint8_t *neighbor_mac_address);
 
 const char *tr_ws_frame(uint8_t frame_type);
